@ajcvickers 2024-11-26
I've looked at this and I don't think I'm ready to approve yet. I think we really need to understand what, "if the final product is itself a delegate, that delegate will itself be in the interpreter mode and therefore less efficient when invoked multiple times when the query runs" means.